Simplify in bound ports configuration panel in vm creation dialog basic panel
