+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-vm/src/main/java/com/microsoft/azure/toolkit/intellij/vm/creation/component/InboundPortRulesForm.java
@@ -7,6 +7,7 @@ package com.microsoft.azure.toolkit.intellij.vm.creation.component;
 import com.microsoft.azure.toolkit.lib.common.form.AzureForm;
 import com.microsoft.azure.toolkit.lib.common.form.AzureFormInput;
 import com.microsoft.azure.toolkit.lib.compute.security.model.SecurityRule;
+import org.apache.commons.collections4.CollectionUtils;
 
 import javax.swing.*;
 import java.awt.event.ActionListener;
@@ -18,10 +19,6 @@ import java.util.Optional;
 import java.util.stream.Stream;
 
 public class InboundPortRulesForm extends JPanel implements AzureForm<List<SecurityRule>> {
-    private JLabel lblPublicInboundPorts;
-    private JPanel pnlPublicInboundsRadios;
-    private JRadioButton rdoAllowSelectedInboundPorts;
-    private JRadioButton rdoNoneInboundPorts;
     private JLabel lblSelectInboundPorts;
     private JPanel pnlInboundPorts;
     private JCheckBox chkHTTP;
@@ -32,15 +29,10 @@ public class InboundPortRulesForm extends JPanel implements AzureForm<List<Secur
 
     public InboundPortRulesForm() {
         super();
-        $$$setupUI$$$();
-        init();
     }
 
     @Override
     public List<SecurityRule> getData() {
-        if (rdoNoneInboundPorts.isSelected()) {
-            return null;
-        }
         final List<SecurityRule> policies = new ArrayList<>();
         Optional.ofNullable(chkHTTP.isSelected() ? SecurityRule.HTTP_RULE : null).ifPresent(policies::add);
         Optional.ofNullable(chkHTTPS.isSelected() ? SecurityRule.HTTPS_RULE : null).ifPresent(policies::add);
@@ -51,15 +43,14 @@ public class InboundPortRulesForm extends JPanel implements AzureForm<List<Secur
 
     @Override
     public void setData(List<SecurityRule> data) {
-        if (data == null) {
-            rdoNoneInboundPorts.setSelected(true);
-        } else {
-            rdoAllowSelectedInboundPorts.setSelected(true);
-            chkHTTP.setSelected(data.contains(SecurityRule.HTTP_RULE));
-            chkHTTPS.setSelected(data.contains(SecurityRule.HTTPS_RULE));
-            chkSSH.setSelected(data.contains(SecurityRule.SSH_RULE));
-            chkRDP.setSelected(data.contains(SecurityRule.RDP_RULE));
+        if (CollectionUtils.isEmpty(data)) {
+            Stream.of(chkRDP, chkHTTP, chkHTTPS, chkSSH).forEach(component -> component.setSelected(false));
+            return;
         }
+        chkHTTP.setSelected(data.contains(SecurityRule.HTTP_RULE));
+        chkHTTPS.setSelected(data.contains(SecurityRule.HTTPS_RULE));
+        chkSSH.setSelected(data.contains(SecurityRule.SSH_RULE));
+        chkRDP.setSelected(data.contains(SecurityRule.RDP_RULE));
     }
 
     @Override
@@ -68,12 +59,11 @@ public class InboundPortRulesForm extends JPanel implements AzureForm<List<Secur
     }
 
     public List<JLabel> getLabels() {
-        return Arrays.asList(lblPublicInboundPorts, lblSelectInboundPorts);
+        return Arrays.asList(lblSelectInboundPorts);
     }
 
     public void addActionListenerToComponents(final ActionListener actionListener) {
-        Stream.of(chkRDP, chkHTTP, chkHTTPS, chkSSH, rdoNoneInboundPorts, rdoAllowSelectedInboundPorts)
-                .forEach(component -> component.addActionListener(actionListener));
+        Stream.of(chkRDP, chkHTTP, chkHTTPS, chkSSH).forEach(component -> component.addActionListener(actionListener));
     }
 
     @Override
@@ -82,25 +72,12 @@ public class InboundPortRulesForm extends JPanel implements AzureForm<List<Secur
         this.pnlRoot.setVisible(flag);
     }
 
-    private void init() {
-        final ButtonGroup inboundPortsGroup = new ButtonGroup();
-        inboundPortsGroup.add(rdoNoneInboundPorts);
-        inboundPortsGroup.add(rdoAllowSelectedInboundPorts);
-        rdoNoneInboundPorts.addItemListener(e -> toggleInboundPortsPolicy(false));
-        rdoAllowSelectedInboundPorts.addItemListener(e -> toggleInboundPortsPolicy(true));
-        rdoNoneInboundPorts.setSelected(true);
-    }
-
-    private void toggleInboundPortsPolicy(boolean allowInboundPorts) {
+    public void toggleInboundPortsPolicy(boolean allowInboundPorts) {
         Stream.of(chkRDP, chkHTTP, chkHTTPS, chkSSH).forEach(chk -> {
             chk.setEnabled(allowInboundPorts);
             if (!allowInboundPorts) {
                 chk.setSelected(false);
             }
         });
-    }
-
-    // CHECKSTYLE IGNORE check FOR NEXT 1 LINES
-    private void $$$setupUI$$$() {
     }
 }
